Move more code into `impl` folder
-  Only keep top-level stuff and public types in the src root

JIRA: FET-608
