--- libs/sdk-ui-pivot/src/impl/base/agTypes.ts
@@ -1,6 +1,0 @@
-// (C) 2007-2021 GoodData Corporation
-import { GridOptions } from "@ag-grid-community/all-modules";
-
-export interface ICustomGridOptions extends GridOptions {
-    enableMenu?: boolean;
-}
