+++ libs/sdk-ui-pivot/src/impl/cell/onCellClick.ts
@@ -1,5 +1,5 @@
 // (C) 2007-2021 GoodData Corporation
-import { TableFacade } from "../../tableFacade";
+import { TableFacade } from "../tableFacade";
 import { ICorePivotTableProps } from "../../publicTypes";
 import { CellEvent } from "@ag-grid-community/all-modules";
 import { invariant } from "ts-invariant";
