+++ libs/sdk-ui-pivot/src/impl/structure/colDefHeaderClass.ts
@@ -1,5 +1,5 @@
 // (C) 2007-2021 GoodData Corporation
-import { TableFacade } from "../../tableFacade";
+import { TableFacade } from "../tableFacade";
 import { ColDef, ColGroupDef } from "@ag-grid-community/all-modules";
 import cx from "classnames";
 import { agColId, ColumnGroupingDescriptorId, isDataColLeaf, isSliceCol } from "./tableDescriptorTypes";
