+++ libs/sdk-ui-pivot/src/impl/structure/headers/AggregationsSubMenu.tsx
@@ -13,7 +13,7 @@ import { IntlShape } from "react-intl";
 import SubMenu from "../../../menu/SubMenu";
 import menuHelper from "./aggregationsMenuHelper";
 import { IColumnTotal } from "./aggregationsMenuTypes";
-import { IMenuAggregationClickConfig } from "../../../privateTypes";
+import { IMenuAggregationClickConfig } from "../../privateTypes";
 
 const MENU_HEADER_OFFSET = -36;
